remove old activity
