+++ app/src/main/AndroidManifest.xml
@@ -17,12 +17,10 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:targetApi="lollipop">
-        <activity android:name=".mvp.MainActivity">
-        </activity>
         <activity
             android:name=".PlayerActivity"
             android:configChanges="orientation|screenSize|keyboardHidden|smallestScreenSize|screenLayout" />
-        <activity android:name=".MainActivityWithNavigation"
+        <activity android:name=".mvp.MainActivity"
             android:theme="@style/Theme.AppCompat.DayNight.DarkActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
